Added the to_seconds() and to_milliseconds() conversion functions.
Disconnect options now using chrono durations.

+++ src/client.cpp
@@ -65,7 +65,7 @@ void client::disconnect()
 	cli_.disconnect()->wait_for_completion(timeout_);
 }
 
-void client::disconnect(long timeout)
+void client::disconnect(int timeout)
 {
 	cli_.disconnect(timeout)->wait_for_completion(timeout_);
 }
